# Conflicts:
#	README.md
#	mkdocs.yml
#	src/OrchardCore.Build/Dependencies.props
#	src/OrchardCore.Build/OrchardCore.Commons.props
#	src/OrchardCore/OrchardCore.Abstractions/Modules/Manifest/ManifestConstants.cs
#	src/OrchardCore/OrchardCore.ContentManagement.Abstractions/Metadata/Builders/ContentPartDefinitionBuilder.cs
#	src/README.md
#	src/docs/getting-started/templates/README.md
#	src/docs/guides/add-admin-menu/README.md
#	src/docs/guides/create-cms-application/README.md
#	src/docs/guides/create-modular-application-mvc/README.md
#	src/docs/guides/decoupled-cms/README.md
#	src/docs/resources/libraries/README.md

+++ src/OrchardCore.Modules/OrchardCore.Search.Elasticsearch/Controllers/AdminController.cs
@@ -4,6 +4,8 @@ using System.Diagnostics;
 using System.Globalization;
 using System.Linq;
 using System.Text.Encodings.Web;
+using System.Text.Json;
+using System.Text.Json.Nodes;
 using System.Threading.Tasks;
 using Fluid;
 using Fluid.Values;
@@ -12,11 +14,12 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Localization;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.AspNetCore.Routing;
+using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Localization;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
-using Newtonsoft.Json;
-using Newtonsoft.Json.Linq;
+using OrchardCore.Admin;
+using OrchardCore.BackgroundJobs;
 using OrchardCore.ContentManagement.Metadata;
 using OrchardCore.DisplayManagement;
 using OrchardCore.DisplayManagement.Notify;
@@ -32,8 +35,11 @@ using YesSql;
 
 namespace OrchardCore.Search.Elasticsearch
 {
+    [Admin("elasticsearch/{action}/{id?}", "Elasticsearch.{action}")]
     public class AdminController : Controller
     {
+        private const string _optionsSearch = "Options.Search";
+
         private readonly ISession _session;
         private readonly ISiteService _siteService;
         private readonly ILiquidTemplateManager _liquidTemplateManager;
@@ -48,6 +54,7 @@ namespace OrchardCore.Search.Elasticsearch
         private readonly INotifier _notifier;
         private readonly ILogger _logger;
         private readonly IOptions<TemplateOptions> _templateOptions;
+        private readonly ElasticConnectionOptions _elasticConnectionOptions;
         private readonly IShapeFactory _shapeFactory;
         private readonly ILocalizationService _localizationService;
 
@@ -69,6 +76,7 @@ namespace OrchardCore.Search.Elasticsearch
             INotifier notifier,
             ILogger<AdminController> logger,
             IOptions<TemplateOptions> templateOptions,
+            IOptions<ElasticConnectionOptions> elasticConnectionOptions,
             IShapeFactory shapeFactory,
             ILocalizationService localizationService,
             IStringLocalizer<AdminController> stringLocalizer,
@@ -88,6 +96,7 @@ namespace OrchardCore.Search.Elasticsearch
             _notifier = notifier;
             _logger = logger;
             _templateOptions = templateOptions;
+            _elasticConnectionOptions = elasticConnectionOptions.Value;
             _shapeFactory = shapeFactory;
             _localizationService = localizationService;
             S = stringLocalizer;
@@ -99,6 +108,11 @@ namespace OrchardCore.Search.Elasticsearch
             if (!await _authorizationService.AuthorizeAsync(User, Permissions.ManageElasticIndexes))
             {
                 return Forbid();
+            }
+
+            if (!_elasticConnectionOptions.FileConfigurationExists())
+            {
+                return NotConfigured();
             }
 
             var indexes = (await _elasticIndexSettingsService.GetSettingsAsync())
@@ -121,12 +135,13 @@ namespace OrchardCore.Search.Elasticsearch
 
             // Maintain previous route data when generating page links.
             var routeData = new RouteData();
-            var pagerShape = await _shapeFactory.CreateAsync("Pager", Arguments.From(new
+
+            if (!string.IsNullOrEmpty(options.Search))
             {
-                pager.Page,
-                pager.PageSize,
-                TotalItemCount = totalIndexes
-            }));
+                routeData.Values.TryAdd(_optionsSearch, options.Search);
+            }
+
+            var pagerShape = await _shapeFactory.PagerAsync(pager, totalIndexes, routeData);
 
             var model = new AdminIndexViewModel
             {
@@ -135,25 +150,26 @@ namespace OrchardCore.Search.Elasticsearch
                 Pager = pagerShape
             };
 
-            model.Options.ContentsBulkAction = [
+            model.Options.ContentsBulkAction =
+            [
                 new SelectListItem(S["Reset"], nameof(ContentsBulkAction.Reset)),
                 new SelectListItem(S["Rebuild"], nameof(ContentsBulkAction.Rebuild)),
                 new SelectListItem(S["Delete"], nameof(ContentsBulkAction.Remove)),
             ];
+
 
             return View(model);
         }
 
         [HttpPost, ActionName(nameof(Index))]
         [FormValueRequired("submit.Filter")]
-        public ActionResult IndexFilterPOST(AdminIndexViewModel model)
-        {
-            return RedirectToAction(nameof(Index), new RouteValueDictionary {
-                { "Options.Search", model.Options.Search }
+        public IActionResult IndexFilterPOST(AdminIndexViewModel model)
+            => RedirectToAction(nameof(Index), new RouteValueDictionary
+            {
+                { _optionsSearch, model.Options.Search }
             });
-        }
 
-        public async Task<ActionResult> Edit(string indexName = null)
+        public async Task<IActionResult> Edit(string indexName = null)
         {
             var IsCreate = string.IsNullOrWhiteSpace(indexName);
             var settings = new ElasticIndexSettings();
@@ -161,6 +177,11 @@ namespace OrchardCore.Search.Elasticsearch
             if (!await _authorizationService.AuthorizeAsync(User, Permissions.ManageElasticIndexes))
             {
                 return Forbid();
+            }
+
+            if (!_elasticConnectionOptions.FileConfigurationExists())
+            {
+                return NotConfigured();
             }
 
             if (!IsCreate)
@@ -195,6 +216,11 @@ namespace OrchardCore.Search.Elasticsearch
             if (!await _authorizationService.AuthorizeAsync(User, Permissions.ManageElasticIndexes))
             {
                 return Forbid();
+            }
+
+            if (!_elasticConnectionOptions.FileConfigurationExists())
+            {
+                return BadRequest();
             }
 
             ValidateModel(model);
@@ -292,13 +318,18 @@ namespace OrchardCore.Search.Elasticsearch
                 return Forbid();
             }
 
+            if (!_elasticConnectionOptions.FileConfigurationExists())
+            {
+                return BadRequest();
+            }
+
             if (!await _elasticIndexManager.ExistsAsync(id))
             {
                 return NotFound();
             }
 
             await _elasticIndexingService.ResetIndexAsync(id);
-            await _elasticIndexingService.ProcessContentItemsAsync(id);
+            await ProcessContentItemsAsync(id);
 
             await _notifier.SuccessAsync(H["Index <em>{0}</em> reset successfully.", id]);
 
@@ -311,6 +342,11 @@ namespace OrchardCore.Search.Elasticsearch
             if (!await _authorizationService.AuthorizeAsync(User, Permissions.ManageElasticIndexes))
             {
                 return Forbid();
+            }
+
+            if (!_elasticConnectionOptions.FileConfigurationExists())
+            {
+                return BadRequest();
             }
 
             if (!await _elasticIndexManager.ExistsAsync(id))
@@ -331,7 +367,7 @@ namespace OrchardCore.Search.Elasticsearch
                 await _elasticIndexSettingsService.UpdateIndexAsync(settings);
             }
 
-            await _elasticIndexingService.ProcessContentItemsAsync(id);
+            await ProcessContentItemsAsync(id);
 
             await _notifier.SuccessAsync(H["Index <em>{0}</em> rebuilt successfully.", id]);
 
@@ -344,6 +380,11 @@ namespace OrchardCore.Search.Elasticsearch
             if (!await _authorizationService.AuthorizeAsync(User, Permissions.ManageElasticIndexes))
             {
                 return Forbid();
+            }
+
+            if (!_elasticConnectionOptions.FileConfigurationExists())
+            {
+                return BadRequest();
             }
 
             if (!await _elasticIndexManager.ExistsAsync(model.IndexName))
@@ -375,6 +416,11 @@ namespace OrchardCore.Search.Elasticsearch
                 return Forbid();
             }
 
+            if (!_elasticConnectionOptions.FileConfigurationExists())
+            {
+                return BadRequest();
+            }
+
             try
             {
                 await _elasticIndexingService.DeleteIndexAsync(model.IndexName);
@@ -393,7 +439,7 @@ namespace OrchardCore.Search.Elasticsearch
         public async Task<IActionResult> Mappings(string indexName)
         {
             var mappings = await _elasticIndexManager.GetIndexMappings(indexName);
-            var formattedJson = JValue.Parse(mappings).ToString(Formatting.Indented);
+            var formattedJson = JNode.Parse(mappings).ToJsonString(System.Text.Json.JOptions.Indented);
             return View(new MappingsViewModel
             {
                 IndexName = _elasticIndexManager.GetFullIndexName(indexName),
@@ -413,12 +459,19 @@ namespace OrchardCore.Search.Elasticsearch
             return RedirectToAction(nameof(Index));
         }
 
-        public Task<IActionResult> Query(string indexName, string query)
-            => Query(new AdminQueryViewModel
+        public async Task<IActionResult> Query(string indexName, string query)
+        {
+            if (!_elasticConnectionOptions.FileConfigurationExists())
+            {
+                return NotConfigured();
+            }
+
+            return await Query(new AdminQueryViewModel
             {
                 IndexName = indexName,
                 DecodedQuery = string.IsNullOrWhiteSpace(query) ? string.Empty : System.Text.Encoding.UTF8.GetString(Convert.FromBase64String(query))
             });
+        }
 
         [HttpPost]
         public async Task<IActionResult> Query(AdminQueryViewModel model)
@@ -426,6 +479,11 @@ namespace OrchardCore.Search.Elasticsearch
             if (!await _authorizationService.AuthorizeAsync(User, Permissions.ManageElasticIndexes))
             {
                 return Forbid();
+            }
+
+            if (!_elasticConnectionOptions.FileConfigurationExists())
+            {
+                return BadRequest();
             }
 
             model.Indices = (await _elasticIndexSettingsService.GetSettingsAsync()).Select(x => x.IndexName).ToArray();
@@ -459,7 +517,7 @@ namespace OrchardCore.Search.Elasticsearch
             var stopwatch = new Stopwatch();
             stopwatch.Start();
 
-            var parameters = JsonConvert.DeserializeObject<Dictionary<string, object>>(model.Parameters);
+            var parameters = JConvert.DeserializeObject<Dictionary<string, object>>(model.Parameters);
             var tokenizedContent = await _liquidTemplateManager.RenderStringAsync(model.DecodedQuery, _javaScriptEncoder, parameters.Select(x => new KeyValuePair<string, FluidValue>(x.Key, FluidValue.Create(x.Value, _templateOptions.Value))));
 
             try
@@ -493,6 +551,11 @@ namespace OrchardCore.Search.Elasticsearch
                 return Forbid();
             }
 
+            if (!_elasticConnectionOptions.FileConfigurationExists())
+            {
+                return BadRequest();
+            }
+
             if (itemIds?.Count() > 0)
             {
                 var elasticIndexSettings = await _elasticIndexSettingsService.GetSettingsAsync();
@@ -518,7 +581,7 @@ namespace OrchardCore.Search.Elasticsearch
                             }
 
                             await _elasticIndexingService.ResetIndexAsync(item.IndexName);
-                            await _elasticIndexingService.ProcessContentItemsAsync(item.IndexName);
+                            await ProcessContentItemsAsync(item.IndexName);
 
                             await _notifier.SuccessAsync(H["Index <em>{0}</em> reset successfully.", item.IndexName]);
                         }
@@ -532,12 +595,13 @@ namespace OrchardCore.Search.Elasticsearch
                             }
 
                             await _elasticIndexingService.RebuildIndexAsync(await _elasticIndexSettingsService.GetSettingsAsync(item.IndexName));
-                            await _elasticIndexingService.ProcessContentItemsAsync(item.IndexName);
+
+                            await ProcessContentItemsAsync(item.IndexName);
                             await _notifier.SuccessAsync(H["Index <em>{0}</em> rebuilt successfully.", item.IndexName]);
                         }
                         break;
                     default:
-                        throw new ArgumentOutOfRangeException(nameof(options.BulkAction), "Unknown bulk action");
+                        return BadRequest();
                 }
             }
 
@@ -563,16 +627,26 @@ namespace OrchardCore.Search.Elasticsearch
 
         private async Task PopulateMenuOptionsAsync(ElasticIndexSettingsViewModel model)
         {
-             var supportedCultures = await _localizationService.GetSupportedCulturesAsync();
+            var supportedCultures = await _localizationService.GetSupportedCulturesAsync();
 
-             model.Cultures = supportedCultures.Select(c => new SelectListItem
-             {
-                 Text = $"{c} ({CultureInfo.GetCultureInfo(c).DisplayName})",
-                 Value = c
-             });
+            model.Cultures = supportedCultures.Select(c => new SelectListItem
+            {
+                Text = $"{c} ({CultureInfo.GetCultureInfo(c).DisplayName})",
+                Value = c
+            });
 
             model.Analyzers = _elasticSearchOptions.Analyzers
                 .Select(x => new SelectListItem { Text = x.Key, Value = x.Key });
         }
+
+        private ViewResult NotConfigured()
+            => View("NotConfigured");
+
+        private static Task ProcessContentItemsAsync(string indexName)
+            => HttpBackgroundJob.ExecuteAfterEndOfRequestAsync("sync-content-items-elasticsearch-" + indexName, async (scope) =>
+            {
+                var indexingService = scope.ServiceProvider.GetRequiredService<ElasticIndexingService>();
+                await indexingService.ProcessContentItemsAsync(indexName);
+            });
     }
 }
